Tree_669
Given a binary search tree and the lowest and highest boundaries as L
and R, trim the tree so that all its elements lies in [L, R] (R >= L).
You might need to change the root of the tree, so the result should
return the new root of the trimmed binary search tree.