Tree_606
You need to construct a string consists of parenthesis and integers
from a binary tree with the preorder traversing way.

The null node needs to be represented by empty parenthesis pair "()".
And you need to omit all the empty parenthesis pairs that don't affect
the one-to-one mapping relationship between the string and the original
binary tree.