HashTable_609
Given a list of directory info including directory path, and all the
files with contents in this directory, you need to find out all the
groups of duplicate files in the file system in terms of their paths.

A group of duplicate files consists of at least two files that have
exactly the same content.