HashTable_508
Given the root of a tree, you are asked to find the most frequent
subtree sum. The subtree sum of a node is defined as the sum of all the
node values formed by the subtree rooted at that node (including the
node itself). So what is the most frequent subtree sum value? If there
is a tie, return all the values with the highest frequency in any order.