DP_689
In a given array nums of positive integers, find three non-overlapping
subarrays with maximum sum.

Each subarray will be of size k, and we want to maximize the sum of all
3*k entries.

Return the result as a list of indices representing the starting
position of each interval (0-indexed). If there are multiple answers,
return the lexicographically smallest one.